restrict dask version?

+++ plantcv/plantcv/hyperspectral/read_data.py
@@ -75,7 +75,7 @@ def read_data(filename):
     dtype_dict = {"1": np.int8, "2": np.int16, "3": np.int32,
                   "4": np.float32, "5": np.float64, "6": np.complex64,
                   "9": np.complex128, "12": np.uint16, "13": np.uint32,
-                  "14": np.uint64, "15": np.float64}
+                  "14": np.uint64, "15": np.uint64}
     header_dict["data type"] = dtype_dict[header_dict["data type"]]
 
     if "default bands" in header_dict:
